+++ crates/node-depman/src/proto.rs
@@ -1,7 +1,10 @@
 use crate::npm_registry::parse_registry_response;
 use crate::package_manager::PackageManager;
 use extism_pdk::*;
-use node_common::{commands, get_globals_dirs, BinField, NodeDistVersion, PackageJson};
+use node_common::{
+    commands::{self, get_global_prefix},
+    BinField, NodeDistVersion, PackageJson,
+};
 use proto_pdk::*;
 use std::collections::HashMap;
 use std::fs;
@@ -310,7 +313,7 @@ pub fn locate_executables(
     };
 
     Ok(Json(LocateExecutablesOutput {
-        globals_lookup_dirs: get_globals_dirs(&env),
+        globals_lookup_dirs: vec!["$PROTO_HOME/tools/node/globals/bin".into()],
         primary: Some(primary),
         secondary,
         ..LocateExecutablesOutput::default()
@@ -321,9 +324,11 @@ pub fn locate_executables(
 pub fn install_global(
     Json(input): Json<InstallGlobalInput>,
 ) -> FnResult<Json<InstallGlobalOutput>> {
+    let env = get_proto_environment()?;
+
     let result = exec_command!(commands::install_global(
         &input.dependency,
-        &input.globals_dir.real_path(),
+        get_global_prefix(&env, input.globals_dir.real_path()),
     ));
 
     Ok(Json(InstallGlobalOutput::from_exec_command(result)))
@@ -333,9 +338,11 @@ pub fn install_global(
 pub fn uninstall_global(
     Json(input): Json<UninstallGlobalInput>,
 ) -> FnResult<Json<UninstallGlobalOutput>> {
+    let env = get_proto_environment()?;
+
     let result = exec_command!(commands::uninstall_global(
         &input.dependency,
-        &input.globals_dir.real_path(),
+        get_global_prefix(&env, input.globals_dir.real_path()),
     ));
 
     Ok(Json(UninstallGlobalOutput::from_exec_command(result)))
@@ -393,7 +400,6 @@ pub fn pre_run(Json(input): Json<RunHook>) -> FnResult<()> {
 
 #[plugin_fn]
 pub fn locate_bins(Json(input): Json<LocateBinsInput>) -> FnResult<Json<LocateBinsOutput>> {
-    let env = get_proto_environment()?;
     let mut bin_path = None;
     let package_path = input.context.tool_dir.join("package.json");
     let manager = PackageManager::detect();
@@ -437,7 +443,7 @@ pub fn locate_bins(Json(input): Json<LocateBinsInput>) -> FnResult<Json<LocateBi
     Ok(Json(LocateBinsOutput {
         bin_path: bin_path.map(PathBuf::from),
         fallback_last_globals_dir: true,
-        globals_lookup_dirs: get_globals_dirs(&env),
+        globals_lookup_dirs: vec!["$PROTO_HOME/tools/node/globals/bin".into()],
         ..LocateBinsOutput::default()
     }))
 }